Added reset and finalize transfer
The transfer action is now trigerred via a notification received when
someone transfer money to the contract account. This will in return
extract the quantity of the transfer and perform rewards computation
and re-dispatch the money to member of the contract.

Also added a `reset` action that clears everything except member -> account
binding.

Fixed also `bindmember` action that was changing the primary as part
of a secondary index update which is not permitted.